[refactor] created new api/supabase folders, separated createClient from queries
