+++ api/supabase/queries/query.ts
@@ -1,0 +1,9 @@
+import supabase from '../createClient';
+
+export default async function queryProjects() {
+  const { data: projects, error } = await supabase.from('Projects').select('*');
+
+  console.log('PROJECTS', projects, 'ERROR', error);
+
+  return { projects, error };
+}
