+++ app/testing/page.tsx
@@ -3,7 +3,7 @@
 import { CSSProperties, useEffect, useState } from 'react';
 import Image from 'next/image';
 import BPLogo from '@/assets/images/bp-logo.png';
-import queryProjects from '../../src/queries/query';
+import queryProjects from '../../api/supabase/queries/query';
 
 interface Project {
   id: string;
@@ -24,7 +24,6 @@ interface Project {
   key_development_milestones: object | null;
 }
 
-// TODO: build some sort of testing page
 export default function Home() {
   const [projects, setProjects] = useState<Project[] | null>(null);
   const [error, setError] = useState<string | null>(null);
